@@ -1,31 +1,42 @@
-apply plugin: 'com.android.application'
+plugins {
+  id 'com.android.application'
+}
 
 android {
-  compileSdkVersion 25
-  buildToolsVersion "25.0.2"
+  compileSdkVersion 34
+
   defaultConfig {
     applicationId "com.xiaopo.flying.stickerview"
-    minSdkVersion 14
-    targetSdkVersion 25
+    minSdkVersion 21
+    targetSdkVersion 34
     versionCode 1
     versionName "1.0"
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+    testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
   }
+
   buildTypes {
     release {
       minifyEnabled false
-      proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+      proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
     }
+  }
+  compileOptions {
+    sourceCompatibility JavaVersion.VERSION_1_8
+    targetCompatibility JavaVersion.VERSION_1_8
   }
 }
 
 dependencies {
-  compile fileTree(dir: 'libs', include: ['*.jar'])
-  androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-    exclude group: 'com.android.support', module: 'support-annotations'
-  })
-  compile "com.android.support:appcompat-v7:$rootProject.ext.supportLibraryVersion"
-  compile "com.github.chrisbanes:PhotoView:$rootProject.ext.photoViewVersion"
-  testCompile 'junit:junit:4.12'
-  compile project(path: ':sticker')
+  implementation fileTree(dir: 'libs', include: ['*.jar'])
+  implementation 'androidx.appcompat:appcompat:1.6.1'
+  implementation "androidx.constraintlayout:constraintlayout:2.1.4"
+  implementation 'com.github.wuapnjie:StickerView:4a82f82fba'
+
+
+  testImplementation 'junit:junit:4.13.2'
+  androidTestImplementation 'androidx.test.ext:junit:1.2.1'
+  androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
+
+  implementation project(path: ':sticker')
 }
