@@ -1,148 +1,36 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 34 // Cập nhật lên SDK phiên bản mới hơn, ví dụ là 33
 
     defaultConfig {
         minSdkVersion 9
-        targetSdkVersion 25
+        targetSdkVersion 34 // Cập nhật target SDK phiên bản mới hơn
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"  // Thay thế cho AndroidX
     }
+
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile "com.android.support:appcompat-v7:$rootProject.ext.supportLibraryVersion"
+    // Thư viện AndroidX thay thế cho android.support
+    implementation "androidx.appcompat:appcompat:1.6.0"  // Thay thế cho appcompat-v7
+    implementation "androidx.constraintlayout:constraintlayout:2.1.4"  // Thay thế cho constraint-layout
+    implementation 'androidx.activity:activity:1.6.0'
+    implementation 'androidx.fragment:fragment:1.5.5'
+
+    // Thư viện hỗ trợ kiểm tra đơn vị
+    implementation "androidx.test.ext:junit:1.2.1"
+    implementation "androidx.test.espresso:espresso-core:3.6.1"
+
+    // Các thư viện khác nếu có
+    compile fileTree(dir: 'libs', include: ['*.jar'])  // Nếu có các thư viện .jar cần tích hợp
 }
-
-//apply plugin: 'com.android.library'
-//apply plugin: 'com.github.dcendents.android-maven'
-//apply plugin: 'com.jfrog.bintray'
-//
-//version = "1.6.0" // #修改# // 这里是aar的版本号
-//
-//android {
-//  compileSdkVersion 25
-//  buildToolsVersion "25.0.2"
-//  resourcePrefix 'stickerview_'
-//
-//  defaultConfig {
-//    minSdkVersion 9
-//    targetSdkVersion 25
-//    versionCode 1
-//    versionName "1.0"
-//
-//    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-//  }
-//  buildTypes {
-//    release {
-//      minifyEnabled false
-//      proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-//    }
-//  }
-//
-//  lintOptions {
-//    abortOnError false
-//  }
-//}
-//
-//dependencies {
-//  compile fileTree(include: ['*.jar'], dir: 'libs')
-//  compile "com.android.support:appcompat-v7:$rootProject.ext.supportLibraryVersion"
-//}
-//
-//def siteUrl = 'https://github.com/wuapnjie/StickerView'
-//// #修改# // 项目的主页地址，我这里是我的PickerView项目在github的链接地址
-//def gitUrl = 'https://github.com/wuapnjie/StickerView.git'
-//// #修改# // 项目 git 地址，我这里同样是用Github上的git地址
-//group =
-//    "com.flying.xiaopo" // #修改# // 组名称，这个相当于依赖的时候 compile 'com.bigkoo:pickerview:1.0.1' “:”号前面的前缀
-//
-//install {
-//  repositories.mavenInstaller {
-//    // This generates POM.xml with proper parameters
-//    pom {
-//      project {
-//        packaging 'aar'
-//        name 'Sticker support For Android' // #修改# // 标题
-//        url siteUrl
-//        // Set your license
-//        licenses {
-//          license {
-//            name 'The Apache Software License, Version 2.0'
-//            url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-//          }
-//        }
-//        developers {
-//          developer {
-//            id 'wuapnjie'
-//            // #修改# // 你的userid，昵称
-//            name 'wupanjie' // #修改# // 用户名
-//            email 'wupanjie0611@gmail.com' // #修改# // 邮箱
-//          }
-//        }
-//        scm {
-//          connection gitUrl
-//          developerConnection gitUrl
-//          url siteUrl
-//        }
-//      }
-//    }
-//  }
-//}
-//
-//task sourcesJar(type: Jar) {
-//  from android.sourceSets.main.java.srcDirs
-//  classifier = 'sources'
-//}
-//
-//task javadoc(type: Javadoc) {
-//  source = android.sourceSets.main.java.srcDirs
-//  classpath += project.files(android.getBootClasspath().join(File.pathSeparator))
-//}
-//
-//task javadocJar(type: Jar, dependsOn: javadoc) {
-//  classifier = 'javadoc'
-//  from javadoc.destinationDir
-//}
-//
-//artifacts {
-//  archives javadocJar
-//  archives sourcesJar
-//}
-//
-//Properties properties = new Properties()
-//properties.load(project.rootProject.file('local.properties').newDataInputStream())
-//bintray {
-//  user = properties.getProperty("bintray.user")
-//  key = properties.getProperty("bintray.apikey")
-//  // 上面两个 user和key 需要留意一下，在local.properites 里面配置的
-//  configurations = ['archives']
-//  pkg {
-//    repo = "maven"
-//    name = "stickerview"
-//    // #修改# //  在 jcenter 上面的项目名字
-//    websiteUrl = siteUrl
-//    vcsUrl = gitUrl
-//    licenses = ["Apache-2.0"]
-//    publish = true
-//  }
-//}
-
-tasks.withType(Javadoc) {
-  options.addStringOption('Xdoclint:none', '-quiet')
-  options.addStringOption('encoding', 'UTF-8')
-  options.addStringOption('charSet', 'UTF-8')
-}
-
