@@ -3,8 +3,9 @@ package com.xiaopo.flying.sticker;
 import android.graphics.Canvas;
 import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.IntRange;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.IntRange;
+import androidx.annotation.NonNull;
 
 /**
  * @author wupanjie
@@ -19,7 +20,8 @@ public class DrawableSticker extends Sticker {
     realBounds = new Rect(0, 0, getWidth(), getHeight());
   }
 
-  @NonNull @Override public Drawable getDrawable() {
+  @NonNull
+  @Override public Drawable getDrawable() {
     return drawable;
   }
 
