@@ -5,10 +5,12 @@ import android.graphics.Matrix;
 import android.graphics.PointF;
 import android.graphics.RectF;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.IntDef;
-import android.support.annotation.IntRange;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.IntDef;
+import androidx.annotation.IntRange;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
@@ -41,7 +43,8 @@ public abstract class Sticker {
     return isFlippedHorizontally;
   }
 
-  @NonNull public Sticker setFlippedHorizontally(boolean flippedHorizontally) {
+  @NonNull
+  public Sticker setFlippedHorizontally(boolean flippedHorizontally) {
     isFlippedHorizontally = flippedHorizontally;
     return this;
   }
