@@ -3,8 +3,10 @@ package com.xiaopo.flying.sticker;
 import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.IntDef;
 import android.view.MotionEvent;
+
+import androidx.annotation.IntDef;
+
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
