@@ -6,7 +6,6 @@ import android.graphics.Bitmap;
 import android.graphics.RectF;
 import android.net.Uri;
 import android.provider.MediaStore;
-import android.support.annotation.NonNull;
 import android.util.Log;
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -14,6 +13,8 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 
 import static java.lang.Math.round;
+
+import androidx.annotation.NonNull;
 
 /**
  * @author wupanjie
